requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
click==8.1.7
cmake==3.31.1
dlib==19.24.2
dlib-bin==19.24.6
dnspython==2.7.0
ecdsa==0.19.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.104.1
h11==0.14.0
httptools==0.6.4
idna==3.10
numpy==1.26.4
opencv-python==4.10.0.84
pillow==11.0.0
pyasn1==0.6.1
pydantic==2.10.3
pydantic_core==2.27.1
pymongo==4.10.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
rsa==4.9
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
scipy==1.14.1